Delete Card Component
The List Component won't register as a susbscriber
to the Card's `onDelete` event when called from the
Card Dialog's `onDelete` event closure for whatever.
Instead, we use a service for component communication.

We will also address the issue of all lists calling
`deleteCard` once we introduce a relationship between
a card and its list in mongo.

In addition, models *shouldn't* be reassigned. If we
delete a Card from a List like how we deleted a List
from a Board, this (especially Dragula) breaks, and
the model gets out of sync. So, we need to update
`this.cards` in place.